requirements.txt
Flask
python-dotenv
slick-pay==0.3
requests